refactor: move contacts actions nearby contacts
