+++ src/views/contacts/actions/tests/restore-contacts.test.tsx
@@ -7,16 +7,16 @@
 import { act } from '@testing-library/react';
 import { times } from 'lodash';
 
-import { useActionRestoreContacts } from './restore-contacts';
-import { UIAction } from './types';
-import { FOLDERS } from '../carbonio-ui-commons/constants/folders';
-import { getFolderIdParts } from '../carbonio-ui-commons/helpers/folders';
-import { populateFoldersStore } from '../carbonio-ui-commons/test/mocks/store/folders';
-import { setupHook, screen } from '../carbonio-ui-commons/test/test-setup';
-import { FOLDERS_DESCRIPTORS, TIMERS } from '../constants/tests';
-import { Contact } from '../legacy/types/contact';
-import { buildContact } from '../tests/model-builder';
-import { getFoldersArray } from '../tests/utils';
+import { UIAction } from '../../../../actions/types';
+import { FOLDERS } from '../../../../carbonio-ui-commons/constants/folders';
+import { getFolderIdParts } from '../../../../carbonio-ui-commons/helpers/folders';
+import { populateFoldersStore } from '../../../../carbonio-ui-commons/test/mocks/store/folders';
+import { setupHook, screen } from '../../../../carbonio-ui-commons/test/test-setup';
+import { FOLDERS_DESCRIPTORS, TIMERS } from '../../../../constants/tests';
+import { Contact } from '../../../../legacy/types/contact';
+import { buildContact } from '../../../../tests/model-builder';
+import { getFoldersArray } from '../../../../tests/utils';
+import { useActionRestoreContacts } from '../restore-contacts';
 
 describe('useActionRestoreContacts', () => {
 	it('should return an object with the specific data', () => {
