+++ src/views/contacts/actions/tests/trash-contacts.test.tsx
@@ -8,14 +8,14 @@ import { faker } from '@faker-js/faker';
 import { act } from '@testing-library/react';
 import { ErrorSoapBodyResponse } from '@zextras/carbonio-shell-ui';
 
-import { useActionTrashContacts } from './trash-contacts';
-import { UIAction } from './types';
-import { FOLDERS } from '../carbonio-ui-commons/constants/folders';
-import { createSoapAPIInterceptor } from '../carbonio-ui-commons/test/mocks/network/msw/create-api-interceptor';
-import { populateFoldersStore } from '../carbonio-ui-commons/test/mocks/store/folders';
-import { screen, setupHook } from '../carbonio-ui-commons/test/test-setup';
-import { Contact } from '../legacy/types/contact';
-import { buildContact } from '../tests/model-builder';
+import { UIAction } from '../../../../actions/types';
+import { FOLDERS } from '../../../../carbonio-ui-commons/constants/folders';
+import { createSoapAPIInterceptor } from '../../../../carbonio-ui-commons/test/mocks/network/msw/create-api-interceptor';
+import { populateFoldersStore } from '../../../../carbonio-ui-commons/test/mocks/store/folders';
+import { screen, setupHook } from '../../../../carbonio-ui-commons/test/test-setup';
+import { Contact } from '../../../../legacy/types/contact';
+import { buildContact } from '../../../../tests/model-builder';
+import { useActionTrashContacts } from '../trash-contacts';
 
 describe("Trash-contacts' actions", () => {
 	it('should return true if the object response is correctly initialized', () => {
